fix: updated rename org setting and Host Name
