+++ src/organizations/components/OrgProfileTab/CopyableLabeledData.tsx
@@ -1,6 +1,7 @@
 // Libraries
 import React, {FC} from 'react'
-import {useDispatch} from 'react-redux'
+import {useDispatch, useSelector} from 'react-redux'
+import {useHistory} from 'react-router-dom'
 
 // Components
 import {
@@ -10,6 +11,8 @@ import {
   HeadingElement,
   FontWeight,
   JustifyContent,
+  Button,
+  IconFont,
 } from '@influxdata/clockface'
 import CopyButton from 'src/shared/components/CopyButton'
 
@@ -19,18 +22,27 @@ import {notify} from 'src/shared/actions/notifications'
 // Utils
 import {copyToClipboardSuccess} from 'src/shared/copy/notifications'
 import 'src/organizations/components/OrgProfileTab/style.scss'
+import {getOrg} from 'src/organizations/selectors'
 
 interface Props {
   id: string
   label: string
   src: string
   name?: string
+  renameableOrg?: boolean
 }
 
-const CopyableLabeledData: FC<Props> = ({id, label, src, name}) => {
+const CopyableLabeledData: FC<Props> = ({id, label, src, name, renameableOrg}) => {
   const dispatch = useDispatch()
+  const org = useSelector(getOrg)
+  const history = useHistory()
+
   const generateCopyText = () => {
     dispatch(notify(copyToClipboardSuccess(label, src)))
+  }
+
+  const handleShowEditOverlay = () => {
+    history.push(`/orgs/${org.id}/about/rename`)
   }
 
   return (
@@ -62,6 +74,15 @@ const CopyableLabeledData: FC<Props> = ({id, label, src, name}) => {
             testID={`copy-btn--${id}`}
             onCopy={generateCopyText}
           />
+          {renameableOrg && 
+            <Button
+              testID="rename-org--button"
+              text="Rename"
+              icon={IconFont.Pencil}
+              onClick={handleShowEditOverlay}
+              size={ComponentSize.ExtraSmall}
+            />
+          }
           {name && (
             <label
               className="code-snippet--label"
